free functions for list struct types
Memory allocated in Java via GraalVM `UnmanagedMemory.calloc` cannot be freed
by libc `free` so the pointers must be passed back to Java code so
`freeUnmanagedMemory.free` can free it.

Here only the array/list structs must be freed because the Java code
allocates only once for a collection of objects.

The array/list struct is placed at the beginning of the allocated chunk
and therefore it's their pointer that must be freed, in turn freeing
every nested item.